separate detail presenter protocol into 3 parts and make typelias for better readability
