+++ Homeworks.xcworkspace/xcuserdata/catarinapolakowsky.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -110,8 +110,8 @@
             filePath = "CarsApp/CarsApp/Modules/CarDetailModule/CarDetailPresenter.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "65"
-            endingLineNumber = "65"
+            startingLineNumber = "53"
+            endingLineNumber = "53"
             landmarkName = "cellForRow(tableView:at:)"
             landmarkType = "7">
             <Locations>
@@ -158,8 +158,8 @@
             filePath = "CarsApp/CarsApp/Modules/CarDetailModule/CarDetailPresenter.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "53"
-            endingLineNumber = "53"
+            startingLineNumber = "41"
+            endingLineNumber = "41"
             landmarkName = "getRowCountInSection(at:)"
             landmarkType = "7">
          </BreakpointContent>
@@ -174,8 +174,8 @@
             filePath = "CarsApp/CarsApp/Modules/CarDetailModule/CarDetailPresenter.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "110"
-            endingLineNumber = "110"
+            startingLineNumber = "98"
+            endingLineNumber = "98"
             landmarkName = "loadCarData()"
             landmarkType = "7">
          </BreakpointContent>
@@ -190,8 +190,8 @@
             filePath = "CarsApp/CarsApp/Modules/CarDetailModule/CarListDetailViewController.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "138"
-            endingLineNumber = "138"
+            startingLineNumber = "139"
+            endingLineNumber = "139"
             landmarkName = "setupActionForDissmissButton()"
             landmarkType = "7">
          </BreakpointContent>
